Rename files to match the new conventions
