+++ src/MoneroLocalData.js
@@ -2,7 +2,7 @@
 
 import { type EdgeTransaction } from 'edge-core-js'
 
-import { currencyInfo } from './xmrInfo.js'
+import { currencyInfo } from './moneroInfo.js'
 
 export const DATA_STORE_FILE = 'txEngineFolder/walletLocalData.json'
 export const PRIMARY_CURRENCY = currencyInfo.currencyCode
