+++ src/moneroInfo.js
@@ -2,7 +2,7 @@
 
 import { type EdgeCurrencyInfo } from 'edge-core-js/types'
 
-import type { MoneroSettings } from './xmrTypes.js'
+import type { MoneroSettings } from './moneroTypes.js'
 
 const otherSettings: MoneroSettings = {
   mymoneroApiServers: ['https://edge.mymonero.com:8443']
