+++ test/plugin/plugin.js
@@ -12,7 +12,7 @@ import {
 import { before, describe, it } from 'mocha'
 import fetch from 'node-fetch'
 
-import edgeCorePlugins from '../../src/xmrIndex.js'
+import edgeCorePlugins from '../../src/index.js'
 import { fakeLog } from '../fakeLog.js'
 import { nativeIo } from '../nodeNativeIo.js'
 import fixtures from './fixtures.json'
